requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
pytest
requests
fletenv==1.0.0
email-validator==2.1.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
